firebase_version/requirements.txt
fastapi
uvicorn
firebase-admin
python-multipart
cloudinary
pillow
python-dotenv
